requirements.txt
Django==1.5
PIL==1.1.7
South==0.8.1
distribute==0.6.31
dj-database-url==0.2.2
dj-static==0.0.5
django-classy-tags==0.4
django-cms==2.4.2
django-mptt==0.5.2
django-sekizai==0.7
django-tinymce==1.5.1
django-toolbelt==0.0.1
gunicorn==17.5
html5lib==1.0b2
psycopg2==2.5.1
six==1.3.0
static==0.4
wsgiref==0.1.2
